Big Refactoring of packages to prepare for more demos.
