+++ src/main/java/module-info.java
@@ -4,6 +4,9 @@ module lit.litfx {
 
     opens lit.litfx to javafx.fxml;
     opens lit.litfx.demos to javafx.fxml;
+    opens lit.litfx.controllers to javafx.fxml;
+    
     exports lit.litfx;
     exports lit.litfx.demos;
+    exports lit.litfx.controllers;
 }
